+++ bitcoin/rpc.py
@@ -43,7 +43,7 @@ except ImportError:
     import urlparse
 
 import bitcoin
-from bitcoin.core import COIN, lx, b2lx, CBlock, CBlockHeader, CTransaction, COutPoint, CTxOut
+from bitcoin.core import COIN, x, lx, b2lx, CBlock, CBlockHeader, CTransaction, COutPoint, CTxOut
 from bitcoin.core.script import CScript
 from bitcoin.wallet import CBitcoinAddress, CBitcoinSecret
 
@@ -318,8 +318,12 @@ class Proxy(BaseProxy):
         """Return hash of best (tip) block in longest block chain."""
         return lx(self._call('getbestblockhash'))
 
-    def getblockheader(self, block_hash):
+    def getblockheader(self, block_hash, verbose=False):
         """Get block header <block_hash>
+
+        verbose - If true a dict is returned with the values returned by
+                  getblockheader that are not in the block header itself
+                  (height, nextblockhash, etc.)
 
         Raises IndexError if block_hash is not valid.
         """
@@ -329,11 +333,22 @@ class Proxy(BaseProxy):
             raise TypeError('%s.getblockheader(): block_hash must be bytes; got %r instance' %
                     (self.__class__.__name__, block_hash.__class__))
         try:
-            r = self._call('getblockheader', block_hash, False)
+            r = self._call('getblockheader', block_hash, verbose)
         except JSONRPCError as ex:
             raise IndexError('%s.getblockheader(): %s (%d)' %
                     (self.__class__.__name__, ex.error['message'], ex.error['code']))
-        return CBlockHeader.deserialize(unhexlify(r))
+
+        if verbose:
+            nextblockhash = None
+            if 'nextblockhash' in r:
+                nextblockhash = lx(r['nextblockhash'])
+            return {'confirmations':r['confirmations'],
+                    'height':r['height'],
+                    'mediantime':r['mediantime'],
+                    'nextblockhash':nextblockhash,
+                    'chainwork':x(r['chainwork'])}
+        else:
+            return CBlockHeader.deserialize(unhexlify(r))
 
 
     def getblock(self, block_hash):